requirements.txt
torch~=2.5.0
torchvision~=0.20.0
tqdm~=4.66.6
scikit-learn~=1.5.2
pillow~=11.0.0
matplotlib~=3.9.2